+++ src/modules/packages.c
@@ -57,6 +57,7 @@ void ffPrintPackages(FFinstance* instance)
         FF_PRINT_PACKAGE(port)
         FF_PRINT_PACKAGE(scoop)
         FF_PRINT_PACKAGE(choco)
+        FF_PRINT_PACKAGE(pkgtool)
 
         //Fix linter warning of unused value of all
         (void) all;